Standardise file parser organisation
