+++ src/file_parsers/rs/mod.rs
@@ -1,21 +1,11 @@
-use crate::file_parsers::{lift::ToSliceParser, shared::NomParser};
+pub mod parser;
 pub mod types;
 
-use anyhow::{Result, anyhow};
-use nom::{
-    Parser,
-    character::complete::{space0, space1, u32 as U},
-    combinator::{all_consuming, opt},
-    multi::many0,
-    sequence::{preceded, terminated},
-};
+use anyhow::Result;
+use parser::parse_rs_str;
 use types::*;
 
-use crate::file_parsers::{
-    FileParser,
-    shared::{quoted_str, unquoted_str, utf16_bom_to_string, version_line},
-    slice::Slice,
-};
+use crate::file_parsers::{FileParser, shared::utf16_bom_to_string};
 
 pub struct RSParser;
 
@@ -27,39 +17,4 @@ impl FileParser for RSParser {
 
         parse_rs_str(&contents)
     }
-}
-
-fn room<'a>() -> impl NomParser<'a, Room> {
-    (
-        opt(terminated(U, space0)),
-        quoted_str,
-        all_consuming(many0(preceded(space1, unquoted_str))),
-    )
-        .map(|(weight, arm_file, rotations)| Room {
-            weight,
-            arm_file,
-            rotations,
-        })
-}
-
-fn parse_rs_str(contents: &str) -> Result<RSFile> {
-    let lines = contents
-        .lines()
-        .map(|l| l.trim())
-        // Skip empty/commented lines
-        .filter(|l| !l.is_empty() && !l.starts_with("//"))
-        .collect::<Vec<_>>();
-    let lines = Slice(lines.as_slice());
-
-    let parser = (
-        version_line().lift(), //
-        many0(room().lift()),
-    )
-        .map(|(version, rooms)| RSFile { version, rooms });
-
-    let (_, room_file) = all_consuming(parser)
-        .parse_complete(lines)
-        .map_err(|e| anyhow!("Failed to parse file: {e:?}"))?;
-
-    Ok(room_file)
 }
